Clih Bot/requirements.txt
gradio
pyautogui
Pillow
opencv-python
cryptography
pyyaml